# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hoteling System API",
    "description": "API for managing desk reservations in a hoteling system",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Initiates Google OAuth login flow",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "description": "Optional URL to redirect after successful login",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logs out the current user and clears authentication cookies",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "description": "Optional URL to redirect after logout",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}